Add a threshold on low-duration Fermi events
We get a lot of Fermi triggers, and many of them are for events that turn out
not to be real GRBs. The false ones all seem to have very low durations, so
we reject things with a trigger duration of less than 0.5 seconds. I can't
find any information online about methods that real researchers use to
automatically triage these events so this kind of simple-minded cut seems to
be as good a cut as anything.